refactor: replace unit conversion logic with official SDK methods for token conversion
