+++ apps/web/src/services/balance-service.ts
@@ -1,6 +1,6 @@
 import { balance } from '@autonomys/auto-consensus';
-import { shannonsToAI3 } from '@/lib/unit-conversion';
 import { getSharedApiConnection } from './api-service';
+import { shannonsToAI3 } from '@/lib/unit-conversions';
 
 export interface WalletBalance {
   free: string;
@@ -13,7 +13,7 @@ export const fetchWalletBalance = async (address: string): Promise<WalletBalance
 
   const balanceData = await balance(api, address);
 
-  // Convert from shannons to AI3
+  // Convert from shannons to AI3 using our wrapper functions
   const freeAI3 = shannonsToAI3(balanceData.free.toString());
   const reservedAI3 = shannonsToAI3(balanceData.reserved.toString());
   const totalAI3 = freeAI3 + reservedAI3;
